Implementation of Variables finished
Variables now are being considered as symbols, and therefore have a consisting
type checking - supported types are bool and integer, as implemented on the
"more operators branch". variables are not defined with a keyword though, this means you
just define it with a "name = value" over a "make a = 10" as I plan to implement
sooner or later.