Evaluator Cleanup
Cleanup of the evaluator.
Also, there was a stupid mistake that made the binder Bind the binary BoundBinaryExpression
falsely, in to put it more precisely set the BoundNodeKind to UnaryExpression. In the short
run this didn't really matter but in the end was a huge mistake that could have cost lives.